Update README.md
In this timed lab, you will be creating an ALU with four operations. This ALU will take in two 8-bit inputs and one 2-bit op code. It will output one 8-bit output. Additionally, you will be given a circuit that you will be tasked to simplify to use the least amount of transistors possible.

2.1 Tasks and Strategy

    Set up your MUX for the ALU output.

    Build your first operation as a subcircuit, and connect it to the MUX.

    Save your work and submit it to Gradescope. This will ensure you get credit for your work even if you do not finish on time.

    Repeat steps 2 and 3 for the other three operations.

    Build your logically equivalent circuit in part B